Move matched class into edit namespace.
