+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Edit/CaretLayer.cs
@@ -17,7 +17,7 @@ using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Utils;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Edit
 {
     public class CaretLayer : CompositeDrawable
     {
